modification

+++ 06d.program.c
@@ -7,7 +7,7 @@ int myaddition(int a,int b);            // Function prototype with return type i
 
 int main()
 {
-    int a,b,sum;
+    float a,b,sum;
     sum=myaddition(5,6);                // Function call
     sum=myaddition(8.9,9);                // Function call
     sum=myaddition(7,8);                // Function call
